Day 7: Added Level select mechanics, LOADING SCREEN and improved UI

+++ Assets/Script/Coin.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class Coin : MonoBehaviour
 {
 
-    [SerializeField] int rotatespeed = 1;
+    [SerializeField] float rotatespeed = 1;
